Moved Parsers to namespace.
This cleans up the main directory, and ensures parser classes are where they are needed.
