+++ src/WpfMath/Exceptions/TexException.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 
 namespace WpfMath.Exceptions
 {
