+++ src/WpfMath/Glue.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using WpfMath.Boxes;
+using WpfMath.Parsers;
 
 namespace WpfMath
 {
