+++ src/WpfMath/Parsers/PredefinedColorParser.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,7 +6,7 @@ using System.Windows.Media;
 using System.Xml.Linq;
 using System.Reflection;
 
-namespace WpfMath
+namespace WpfMath.Parsers
 {
     internal class PredefinedColorParser
     {
