+++ src/WpfMath/Parsers/PredefinedTexFormulaSettingsParser.cs
@@ -1,0 +1,79 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Xml.Linq;
+
+namespace WpfMath.Parsers
+{
+    /// <summary>
+    /// Parses settings for predefined formulas from XML file.
+    /// </summary>
+    internal class TexPredefinedFormulaSettingsParser
+    {
+        private static readonly string resourceName = TexUtilities.ResourcesDataDirectory + "TexFormulaSettings.xml";
+
+        static TexPredefinedFormulaSettingsParser()
+        {
+        }
+
+        private static void AddToMap(IEnumerable<XElement> mapList, string[] table)
+        {
+            foreach (var map in mapList)
+            {
+                var character = map.AttributeValue("char");
+                var symbol = map.AttributeValue("symbol");
+                Debug.Assert(character != null);
+                Debug.Assert(symbol != null);
+                Debug.Assert(character.Length == 1);
+                table[character[0]] = symbol;
+            }
+        }
+
+        private XElement rootElement;
+
+        public TexPredefinedFormulaSettingsParser()
+        {
+            var doc = XDocument.Load(new System.IO.StreamReader(Assembly.GetExecutingAssembly().GetManifestResourceStream(resourceName)));
+            this.rootElement = doc.Root;
+        }
+
+        public IList<string> GetSymbolMappings()
+        {
+            var mappings = new string[TexFontInfo.charCodesCount];
+            var charToSymbol = rootElement.Element("CharacterToSymbolMappings");
+            if (charToSymbol != null)
+                AddToMap(charToSymbol.Elements("Map"), mappings);
+            return mappings;
+        }
+
+        public IList<string> GetDelimiterMappings()
+        {
+            var mappings = new string[TexFontInfo.charCodesCount];
+            var charToDelimiter = rootElement.Element("CharacterToDelimiterMappings");
+            if (charToDelimiter != null)
+                AddToMap(charToDelimiter.Elements("Map"), mappings);
+            return mappings;
+        }
+
+        public HashSet<string> GetTextStyles()
+        {
+            var result = new HashSet<string>();
+
+            var textStyles = rootElement.Element("TextStyles");
+            if (textStyles != null)
+            {
+                foreach (var textStyleElement in textStyles.Elements("TextStyle"))
+                {
+                    var name = textStyleElement.AttributeValue("name");
+                    Debug.Assert(name != null);
+                    result.Add(name);
+                }
+            }
+
+            return result;
+        }
+    }
+}
